+++ Tests/Desktop/Network/AsyncDicomCGetProviderTests.cs
@@ -126,6 +126,12 @@ namespace Dicom.Network
         {
         }
 
+#if NETSTANDARD2_1 || NETCOREAPP3_0 || NETCOREAPP3_1
+        public async IAsyncEnumerable<DicomCGetResponse> OnCGetRequestAsync(DicomCGetRequest request)
+        {
+            yield return new DicomCGetResponse(request, DicomStatus.Success);
+        }
+#else
         public async Task<IEnumerable<Task<DicomCGetResponse>>> OnCGetRequestAsync(DicomCGetRequest request)
         {
             return InnerOnCGetRequestAsync();
@@ -135,6 +141,7 @@ namespace Dicom.Network
                 yield return Task.FromResult(new DicomCGetResponse(request, DicomStatus.Success));
             }
         }
+#endif
     }
 
     public class PendingAsyncDicomCGetProvider : DicomService, IDicomServiceProvider, IAsyncDicomCGetProvider
@@ -171,6 +178,14 @@ namespace Dicom.Network
         {
         }
 
+#if NETSTANDARD2_1 || NETCOREAPP3_0 || NETCOREAPP3_1
+        public async IAsyncEnumerable<DicomCGetResponse> OnCGetRequestAsync(DicomCGetRequest request)
+        {
+            yield return new DicomCGetResponse(request, DicomStatus.Pending);
+            yield return new DicomCGetResponse(request, DicomStatus.Pending);
+            yield return new DicomCGetResponse(request, DicomStatus.Success);
+        }
+#else
         public async Task<IEnumerable<Task<DicomCGetResponse>>> OnCGetRequestAsync(DicomCGetRequest request)
         {
             return InnerOnCGetRequestAsync();
@@ -182,6 +197,7 @@ namespace Dicom.Network
                 yield return Task.FromResult(new DicomCGetResponse(request, DicomStatus.Success));
             }
         }
+#endif
     }